feat(briefing): aggregator inversion + shim reduction (commit 9/11)
Workstream 6 Step 1 commit 9 — architectural inversion. The 802-line
briefing-service.js orchestrator splits into:

  - briefing-aggregator.js (NEW, 710 lines) — owns ALL orchestration:
    generateAndStoreBriefing, generateBriefingInternal, getBriefingBySnapshotId,
    refreshEventsInBriefing, refreshTrafficInBriefing, refreshNewsInBriefing,
    getOrGenerateBriefing, plus the inFlightBriefings concurrency Map.
  - briefing-service.js (38 lines) — pure re-export facade. ALL function
    bodies preserved byte-for-byte; only the import section is rewritten.

CUMULATIVE briefing-service.js shrinkage across Commits 4-9:
  3,683 → 38 lines (-3,645 net, ~99.0% reduction from session start).

============================================================
TWO MASTER ARCHITECT REVISIONS APPLIED
============================================================

Revision 1 (Facade Purity): the shim re-exports all 17 symbols, NOT just
the 11 currently-exported ones. The 6 discoverX pipeline contracts
(discoverSchools, discoverWeather, discoverAirport, discoverNews,
discoverTraffic, discoverEvents) are now re-exported from
briefing-service.js even though they were never previously exported.
Rationale: defensive coverage for any unseen test harness or debug
script that might import discoverX from briefing-service.js. Recon
confirmed zero such callers exist today, but the safety belt is cheap.

Revision 2 (Memory Chain): claude_memory row #299 chained from #298
(Commit 8.5) — strict chronological execution trail. NOT rebased to
#294 (Commit 4) like prior memories.

============================================================
ESM BINDING-IDENTITY VERIFICATION (load-bearing check)
============================================================
The inFlightBriefings Map is module-private state in briefing-aggregator.js.
ESM module-singleton semantics guarantee one Map per Node process. But
this only works if ALL callers reach the orchestrator through the same
module identity.

Verified via dynamic-import test:
  shim.generateAndStoreBriefing === aggregator.generateAndStoreBriefing
  → TRUE (4-out-of-4 identity checks pass)

This proves ESM re-exports preserve binding identity. A caller importing
generateAndStoreBriefing from the shim and another caller importing it
directly from the aggregator hit the SAME function reference and
therefore the SAME inFlightBriefings Map.

============================================================
ORPHAN IMPORT CLEANUP
============================================================
Removed from briefing-service.js (now only used by extracted code):
  - z (zod), callModel, safeJsonParse, getMarketForLocation, matrixLog,
    isTrafficStale, validateEventsHard, needsReadTimeValidation,
    VALIDATION_SCHEMA_VERSION, normalizeEvent, generateEventHash,
    deduplicateEventsSemantic, findOrCreateVenue, lookupVenue,
    geocodeEventAddress, searchPlaceWithTextSearch, validateVenueAddress,
    deactivatePastEvents, getTomTomTraffic, fetchRawTraffic,
    haversineDistanceMiles, sendModelErrorAlert, dumpLastBriefingRow,
    briefingLog, OP, writeSectionAndNotify, CHANNELS, errorMarker.
  - All schema imports except via aggregator's own focused list.
  - All drizzle-orm helpers — moved to aggregator.
  - GOOGLE_MAPS_API_KEY const — orphan since Commit 4; deleted (not moved).
  - inFlightBriefings Map — moved to aggregator.

============================================================
SCOPE DISCIPLINE
============================================================
Per Master Architect SCOPE CONSTRAINT: pure structural extraction +
orphan-import cleanup + ONE additive surface change (Revision 1's
discoverX re-exports). NO behavioral changes to:
  - generateAndStoreBriefing dedup logic
  - Advisory lock acquisition / release pattern
  - Promise.allSettled fan-out across 5 pipelines
  - Schools cache lookup
  - Final-assembly fallbacks
  - Bulk-failure path
  - getOrGenerateBriefing freshness gating logic
All function bodies copied byte-for-byte.

VERIFICATION:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on both files: pass
  - Dynamic-import identity test: 4/4 TRUE
  - Shim has zero local function definitions (pure re-export facade)
  - All 17 expected exports present in shim:
    deduplicateEvents, discoverAirport, discoverEvents, discoverNews,
    discoverSchools, discoverTraffic, discoverWeather,
    fetchEventsForBriefing, fetchRideshareNews, fetchSchoolClosures,
    fetchTrafficConditions, fetchWeatherConditions, filterInvalidEvents,
    generateAndStoreBriefing, getBriefingBySnapshotId,
    getOrGenerateBriefing, refreshEventsInBriefing
  - 4 external callers verified resolving via shim:
    server/api/briefing/briefing.js, server/api/location/snapshot.js,
    server/lib/briefing/README.md, server/api/briefing/README.md

CLAUDE_MEMORY: row #299 logged with parent_id=298 (Revision 2).

Per PLAN_workstream6_remaining-2026-05-02.md §3.9 — Phase 1 shim with
all public symbols re-exported so external callers don't migrate yet.
Phase 2 (post-soak ≥7 days from prod deploy) will migrate callers to
direct imports from briefing-aggregator.js + pipelines/*.js, after
which this shim can be deleted.
